+++ lib/app/data/storage/import/import_transactions_from_text.dart
@@ -38,7 +38,7 @@ void showImportTransactionsFromTextInput(
               if (!singleTransactionInput.exist) {
                 final t = createNewTransactionFromDateDescriptionAmount(
                   account,
-                  singleTransactionInput.date.asDate(),
+                  singleTransactionInput.date.asDate() ?? DateTime.now(),
                   singleTransactionInput.description.asString(),
                   singleTransactionInput.amount.asAmount(),
                 );